Move generated code into inst include, rename directories
