+++ inst/cpp_generation/model_output.hpp.in
